Created Spring Boot remote server and created client functions library for interacting with cloud server

See merge request s4adatia/cs-346-project!33

+++ console/src/test/kotlin/taskquest/console/views/MainTest.kt
@@ -5,10 +5,15 @@ import java.io.File
 
 internal class MainTest {
     private fun mainTestFunc(params: Array<String>) {
-        File("data.json").copyTo(File("backup.json"), true)
-        File("data.json").delete()
+        if (File("data.json").exists()) {
+            File("data.json").copyTo(File("backup.json"), true)
+            File("data.json").delete()
+        }
         main(params)
-        File("backup.json").copyTo(File("data.json"), true)
+        if (File("backup.json").exists()) {
+            File("backup.json").copyTo(File("data.json"), true)
+            File("backup.json").delete()
+        }
     }
 
     @Test
